Fix a leak and close the connection of a request that timed out.
This fixes a four year old bug; when a request times out, then
setting CURLM_STATE_COMPLETED here, followed by a break, causes
the code to continue at the end of the function where the
message is sent to the user with multi_addmsg() after which
the state is set to CURLM_STATE_MSGSENT.

An easy handle in that state never gets out of that state anymore:
the state does nothing. Cleaning up the easy handle from the user
application (after seeing the request is CURLMSG_DONE) doesn't
change that fact. As a result, the easy handle gets stuck, owning
a connection set to 'close' that is never closed or cleaned up.

This commit set to the state to CURLM_STATE_DONE and does NOT break
out of the loop. As a result the easy handle is cleaned up as a
handle with an error, as it should. It's connection is closed.

However, problems remain after this:
1) Because this was (likely) the last connection, the bundle is deleted
   as well now; which results in libcurl forgetting that this server
   can do pipelining. The new connection that is created for subsequent
   requests, or for handles in the pipeline after this one that didn't
   time out cause new connections to be created until we finally
   receive headers for one of them. This is a random process and we
   can end up any arbitrary number of connections; which afterwards
   will NOT desolve itself (they will all kept being used).
2) The requests that were still in the pipeline will attempt a new
   connection - then discover they have timed out and ALSO call
   connclose, causing the JUST created fresh connection to be closed
   too. The result is a rapid brust of calls to connect() and close().

Both problems are addressed in subsequent commits because they are
exclusively pipeline related, while this commit isn't.