pipelining: Fix connection handling under timeouts.
Data corruption (1420):

This fixes a four year old bug in multi_runsingle(); when a request
times out and then setting CURLM_STATE_COMPLETED, followed by a break
causes the code to continue at the end of the function where the
message is sent to the user with multi_addmsg() after which the state
is set to CURLM_STATE_MSGSENT.

An easy handle in that state never gets out of that state anymore:
the state does nothing. Cleaning up the easy handle from the user
application (after seeing the request is CURLMSG_DONE) doesn't
change that fact. As a result, the easy handle gets stuck, owning
a connection set to 'close' that is never closed or cleaned up.

This commit sets the state to CURLM_STATE_DONE and does NOT break
out of the loop. As a result the easy handle is cleaned up as a
handle with an error, as it should. It's connection is closed.

This commit fixes https://sourceforge.net/p/curl/bugs/1420/
by stopping new requests from being added to such connections
(the changes in ConnectionExists, but actually probably sufficient
that now we close and cleanup the connection) causing their replies
to arrive out of sync.

--
Connection flooding (comments from me in 1420):

While pipelining it is possible that a request never reached the DO
state but still timed out because a request in the pipeline before
it took a long time.

The logic of curl is such that first a new connection is created
and only then the handles on it are checked for timeouts, hence
it is possible that a handle times out without ever having sent
a GET request over their (current) connection. In that case we don't
want to close that connection: it is a healthy connection that probably
was JUST created as replacement for the broken pipe at which the
timeout really occurred.

Note that the 'bits.close' bit is set by default upon construction of
a connectdata struct only to be reset (when the user specifically
requests so) AFTER the connection is established. We can therefore
not use that bit to test if a connection timed out (which also sets
that bit) because the close bit is also set BEFORE it even connects
and when there is a delay between the call to connect() and the
connection being established many requests can be readded before
the first establishes a connection.

This commit fixes both problems by using a new bit (timedout) that is
FALSE upon creation and is only set when the connection times out and
then use this bit to decide if connection should be closed (or can be
reused).

As a result of this commit (using my testsuite available at
https://github.com/CarloWood/curl_http_pipeline_tester) only a single
active connection is created and used when a full pipe breaks due to
a timeout.

That is, although this commit fixes the burst of connect()/close()
calls, the problem remains that the server bundle was refreshed,
so curl forgot about this server being able to do pipelining...
This is fixed in a next commit.