Only assign NULL to data->easy_conn directly.
This commits replaces occurances of *connp = NULL with
data->easy_conn = NULL, adding DEBUGASSERT to make sure
that this is allowed.

What this comes down to is that we only want to set easy_conn
to NULL when the connection is owned by the SessionHandle
that contains that easy_conn. Thus:

  data->easy_conn = NULL;

is only allowed while

  data->easy_conn->data == data

This is a preparation for the next commit that will need this.

Note that this commit also removes a duplicated variable
in curl_multi_remove_handle (nl. 'easy') because it was
an alias for 'data' and having two different variables for
the exact same thing did not make the code better understandable.
In particular, now we can read 'data->easy_conn->data == data'
and 'data->easy_conn->data = data' which matches the above ;)

Finally, in Curl_reconnect_request() we pass connp to
Curl_done() instead of &conn (where conn is a local variable).
The reason is that 'connp' fullfills the above assertion that
Curl_done needs while '&conn' doesn't. There is not functional
difference since 'conn' is not used below this call, *connp
is assigned NULL below that call, and initially conn == *connp.