curl_multi_remove_handle: correctly close conn
When an easy handle is removed premature (before we processed (all of)
the reply), and it already has sent data to the server then that
connection must be closed. Also when there is no pipelining involved.

easy_owns_conn is now set later, because the ownership is changed
and it must reflect that in order to get the connection closed.